adding token to book, customer and transaction

+++ controllers/book-controller.js
@@ -1,45 +1,62 @@
 const dbBook = require('../models/book-model');
+const jwt = require('jsonwebtoken');
 
 var createBook = function(req, res) {
-  dbBook.create(req.body, (err, docs) => {
-    if (err) {
-      res.send(err)
-    } else {
-      res.send(docs)
-    }
-  })
+  if (req.decoded.role == 'admin') {
+    dbBook.create(req.body, (err, docs) => {
+      if (err) {
+        res.send(err)
+      } else {
+        res.send(docs)
+      }
+    })
+  } else {
+    res.send('Unauthorized')
+  }
 }
 
 
 var getAllBook = function(req, res) {
-  dbBook.find({}, (err, docs) => {
-    if (err) {
-      res.send(err)
-    } else {
-      res.send(docs)
-    }
-  })
+  if (req.decoded.role == 'admin' || req.decoded.role == 'client') {
+    dbBook.find({}, (err, docs) => {
+      if (err) {
+        res.send(err)
+      } else {
+        res.send(docs)
+      }
+    })
+  } else {
+    res.send('Unauthorized')
+  }
 }
 
 
 var updateBook = function(req, res) {
-  dbBook.findByIdAndUpdate(req.params.id, req.body, (err,docs) => {
-    if (err) {
-      res.send(err)
-    }else {
-      res.send(docs)
-    }
-  })
+  if (req.decoded.role == 'admin') {
+    dbBook.findByIdAndUpdate(req.params.id, req.body, (err,docs) => {
+      if (err) {
+        res.send(err)
+      }else {
+        res.send(docs)
+      }
+    })
+  } else {
+    res.send('Unauthorized')
+  }
 }
 
 var deleteBook = function(req, res) {
-  dbBook.findByIdAndRemove(req.params.id, (err, docs) => {
-    if (err) {
-      res.send(err)
-    } else {
-      res.send(docs)
-    }
-  })
+  if (req.decoded.role == 'admin') {
+    dbBook.findByIdAndRemove(req.params.id, (err, docs) => {
+      if (err) {
+        res.send(err)
+      } else {
+        res.send(docs)
+      }
+    })
+  } else {
+    res.send('Unauthorized')
+  }
 }
 
 module.exports = {
